tp1/tp1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyserial==3.5
python-multipart==0.0.5